pull out App component
